server/requirements.txt
asgiref==3.6.0
attrs==21.3.0
awscrt==0.16.9
boto3==1.26.129
botocore==1.29.129
certifi==2022.12.7
charset-normalizer==3.1.0
chirptext==0.1.2
dj-database-url==0.5.0
Django==4.1.7
django-cors-headers==3.10.1
django-environ==0.10.0
django-on-heroku==1.1.2
django-polymorphic==3.1.0
django-rest-passwordreset==1.2.1
django-rest-polymorphic==0.1.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.4
jamdict==0.1a11.post2
jamdict-data==1.5
jmespath==1.0.1
jsonfield==3.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
puchikarui==0.1
PyJWT==2.3.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2021.3
requests==2.29.0
romkan==0.2.1
s3transfer==0.6.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==4.2.0
supermemo2==2.0.0
urllib3==1.26.15
whitenoise==6.0.0
Authlib~=1.0.0
